+++ src/library/examples/cursor.rs
@@ -15,7 +15,7 @@ fn main() -> Result<(), grebedb::Error> {
 
     println!("Printing all the key-values...");
 
-    for (key, value) in db.cursor() {
+    for (key, value) in db.cursor()? {
         println!(
             "Cursor key = {}, value = {}",
             std::str::from_utf8(&key).unwrap(),
@@ -25,7 +25,7 @@ fn main() -> Result<(), grebedb::Error> {
 
     println!("Printing all the key-values starting from [key:0004, key:0008) ...");
 
-    let cursor = db.cursor_range(Some("key:0004"), Some("key:0008"))?;
+    let cursor = db.cursor_range("key:0004".."key:0008")?;
 
     for (key, value) in cursor {
         println!(