+++ src/tool/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 (This log only contains changes for grebedb-tool. Changes for the library are located in its own file.)
 
+## 0.2.0 (Unreleased)
+
+* Updated dependency to grebedb 0.2.0
+
 ## 0.1.0 (2021-03-28)
 
 Initial version.